add rsdi as a dependency for DI implementation
