+++ scraper/package.json
@@ -14,6 +14,7 @@
         "debug": "^4.3.1",
         "dotenv": "^8.2.0",
         "got": "^11.8.2",
+        "rsdi": "^1.0.11",
         "typescript": "^4.2.4"
     },
     "devDependencies": {
